+++ acme/services/Configuration.py
@@ -95,7 +95,6 @@ class Configuration(object):
 				'cse.maxExpirationDelta'				: config.getint('cse', 'maxExpirationDelta',			fallback=60*60*24*365*5),	# 5 years, in seconds
 				'cse.requestExpirationDelta'			: config.getfloat('cse', 'requestExpirationDelta',		fallback=10.0),	# 10 seconds
 				'cse.originator'						: config.get('cse', 'originator',						fallback='CAdmin'),
-				'cse.enableNotifications'				: config.getboolean('cse', 'enableNotifications', 		fallback=True),
 				'cse.enableRemoteCSE'					: config.getboolean('cse', 'enableRemoteCSE', 			fallback=True),
 				'cse.enableTransitRequests'				: config.getboolean('cse', 'enableTransitRequests',		fallback=True),
 				'cse.enableValidation'					: config.getboolean('cse', 'enableValidation', 			fallback=True),