+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [RUNTIME] Due to the restructuring of the project structure the CSE must now be started like this: ```python3 -m acme```
 
 ### Removed
+- [CSE] Removed "cse.enableNotifications" configuration option. Notifications are now always enabled.
 - [APPS] Removed the example AEs from the CSE. This makes the CSE a bit smaller and also removes a big dependency to the non-portable psutils package. They will be available in  separate project in the future.
 - [HTTP] Removed the *server.http.multithreaded* configuration option. The http server now always runs in multithread mode.